requirements.txt
bcrypt==4.1.3
beautifulsoup4==4.12.3
cairocffi==1.7.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
cssselect2==0.7.0
customtkinter==5.2.2
darkdetect==0.8.0
defusedxml==0.7.1
docopt==0.6.2
Faker==26.0.0
freetype-py==2.4.0
idna==3.7
Js2Py==0.74
more-itertools==10.2.0
numpy==2.0.0
packaging==24.0
pillow==10.3.0
pycairo==1.26.0
pycparser==2.22
pyjsparser==2.7.1
pyperclip==1.9.0
PyPrind==2.11.3
pySmartDL==1.3.4
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.32.3
scipy==1.13.1
setuptools==70.1.0
six==1.16.0
soupsieve==2.5
svgwrite==1.4.3
tinycss2==1.3.0
typeguard==4.3.0
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
webencodings==0.5.1
